change the upwind example and make it as a test
